Fix namespace typo UseCasese -> UseCases
